Refactor api code in index.ts
The code is not broken into separate logical modules under api.ts
