+++ src/api/index.ts
@@ -1,0 +1,4 @@
+import api from './api';
+export { defaultPort } from './api';
+
+export default api;
